This is my Project 1 Script
What this script does is give you the time with a "good evening" message and opens google chrome for you. The reason it says good evening is because I couldn't get the loop working to make it say different things throughout different times of the day, so i just went with what it would say at the time of creation. Hope you like it!

To get the script to open the program, i went to quora of all places LOL. Here is the cited link https://www.quora.com/How-do-I-open-a-web-browser-using-powershell